Modified songList to create it's own song object. Now pushSong works with strings

+++ song.h
@@ -13,6 +13,6 @@ public:
 public:
     void setSong(std::string song) { this->songTittle = song; };
     std::string getSong() { return this->songTittle; };
-    int getSizeOf() { return songTittle.size(); };
+    size_t getSizeOf() { return songTittle.size(); };
 
 };
